model - enhancements to error reporting
If a template was missing from the DOM but declared in an extension, the
error messaging only displayed in the appPreView (which quickly
disappeared) so it now displays in the appPreView (in case a load error
occurs) and in the view as well.  In addition, it's reported to the
console as an error.