admin syndication - updates to ebay ui
ebay profile editor now uses panel approach. panels can be closed and
are persistent between profiles and sessions.
create profile is not just a profile id and, on create, opens the
profile in edit mode.
added a button for refresh listings. does NOT work yet.
moved a lot of code around to make all this work.