admin syndication - enhancement to template editor
Added a new button to the html toolbar that will open the list of files
associated with the profile. Clicking the file will add it to the html.

Also updated the add new file icon and added an additional call to the
file upload success so that it will automatically refresh the file
content list for the profile so that going back into the file list
display will show the most up to date contents.